wip: builds noop wasm module
