+++ internal/rust/parser/scanner/src/scanner.rl
